Former-commit-id: 63cc3f7c6c23901ecf81b24335d80a22d321c55b

+++ src/gpt.py
@@ -35,10 +35,14 @@ def call_gpt(operator, source, target, options):
     # ------------------------------------#
     # Building gpt command line #
     # ------------------------------------#
+    # Create absolute target and source paths 
+    source = os.path.abspath(source)
+    target = os.path.abspath(target)
+
     gpt_options = ' '.join([parameter(key, value) for key, value in options.items() if value is not None])
     targetopt = ("-t \"%s\"" % target if target else "")
     gpt_command = "gpt %s -f GeoTIFF %s -Ssource=\"%s\" %s" % (operator, targetopt, source, gpt_options)
-    
+
     # ------------------------------------#
     #    Run gpt command line   #
     # ------------------------------------#
@@ -46,6 +50,8 @@ def call_gpt(operator, source, target, options):
     try:
         process = subprocess.Popen(gpt_command,
                                    shell=True,
+                                   #executable='/bin/bash',
+                                   #cwd=running_dir,
                                    stdin=subprocess.PIPE,
                                    stdout=subprocess.PIPE,
                                    stderr=subprocess.PIPE, )
@@ -56,10 +62,10 @@ def call_gpt(operator, source, target, options):
             raise Exception(stderr)  # or  if process.returncode:
         if 'Error' in stdout:
             raise Exception()
-    except Exception, message:
+    except Exception as message:
         print(str(message))
         # sys.exit(1)  # or sys.exit(process.returncode)
     
     # Change output name (dummy resolution to solve SNAP bug of automatically adding extension)
-    os.rename(target + ".tif", target)
-    
+    if os.path.exists(target + ".tif"):
+        os.rename(target + ".tif", target)
